Support non fixture start
